# Replace text/template for a significant log speedup