# Summary

<!-- Describe your changes in detail here, if it closes an open issue, include "Closes #<issue>" -->
Replaces `text/template` with a simple `strings.Builder` for a pretty dramatic performance improvement:

```plaintext
goos: darwin
goarch: arm64
pkg: github.com/FollowTheProcess/actions/log
cpu: Apple M1 Pro
      │  before.txt  │              after.txt              │
      │    sec/op    │   sec/op     vs base                │
Log-8   6066.5n ± 0%   357.2n ± 0%  -94.11% (p=0.000 n=30)

      │ before.txt  │             after.txt              │
      │    B/op     │    B/op     vs base                │
Log-8   2216.0 ± 0%   320.0 ± 0%  -85.56% (p=0.000 n=30)

      │ before.txt │             after.txt              │
      │ allocs/op  │ allocs/op   vs base                │
Log-8   87.00 ± 0%   13.00 ± 0%  -85.06% (p=0.000 n=30)
```